==> newFlowCohortExtractionPipeline/BIDMC/icd_neg/1_BIDMC_icd_filter.ipynb <==
import pandas as pd
from thunderpack import ThunderReader
from tqdm import tqdm
import re
import random

reader = ThunderReader('/home/jsearle/bigDrive/Dropbox/zz_EHR_Thunderpacks/BIDMC/thunderpack_icd_9_10_nax_1m_BIDMC')
key_length = len(list(reader.keys()))
print(key_length)

# filter by below icd codes
code_regex = "^(?!I60|430).*"
dfs = []
for i in tqdm(range(1, key_length + 1)):
    df = reader[f'ICD_partition_{i}']
    df = df[df['DiagnosisCodeWithDots'].astype(str).str.match(code_regex, flags = re.I)]
    dfs.append(df)

# show num of total SAH icd codes given
# show example of df
filtered_icd_df = pd.concat(dfs, axis=0, ignore_index=True)
print(len(filtered_icd_df))
filtered_icd_df.head()

# clean up df, keep only relevant info
keepColumns = ['BDSPPatientID', 'AdmissionDate', 'DiagnosisCodeWithDots']
clean_icd_df = filtered_icd_df[keepColumns]
clean_icd_df.head()

# rename columns
rename_dict = { 
    'DiagnosisCodeWithDots': 'ICD', 
}

clean_icd_df = clean_icd_df.rename(columns=rename_dict)

clean_icd_df.head()


# RANDOMIZE order of df, 25 seed for reproducability
clean_icd_df = clean_icd_df.sample(frac=1, random_state=25).reset_index(drop=True)
clean_icd_df.head()

#filter to only include unique patients
unique_patients_icd_df = clean_icd_df.drop_duplicates(subset='BDSPPatientID', keep='first')
print(len(unique_patients_icd_df))

# Grab the first 50000 (already randomized) to make the df easier to work with
icd_neg_df = unique_patients_icd_df.head(50000)
print(len(icd_neg_df))

# save df as csv
icd_neg_df.to_csv('1_icd_neg_df.csv', index=False)
