==> newFlowCohortExtractionPipeline/BIDMC/icd_neg/2_BIDMC_admission_filter copy.ipynb <==
import pandas as pd
from thunderpack import ThunderReader
from tqdm import tqdm
import os
import re

# get note metadata for each year
# filter by discharge summary type
startYear = 2010
dfs = []
for i in tqdm(range(0, 15)):
    df = pd.read_csv(f'/home/jsearle/bigDrive/Dropbox/zz_EHR_Thunderpacks/BIDMC/BIDMC_Deidentified_Notes_March14th2024/bidmc_notes_{startYear + i}_metadata.csv')
    df = df[(df['NoteTypeFull'] == 'Initial Note') | (df['NoteTypeFull'] == 'Initial note')]
    dfs.append(df)

# show dataframe, number of admission notes

admission_notes_df = pd.concat(dfs, axis=0, ignore_index=True)
numAdmissionNotes = len(admission_notes_df)
print(numAdmissionNotes)
admission_notes_df.tail()

# clean up df, keep only relevant info
keepColumns = ['BDSPPatientID', 'CreateDate',  'DeidentifiedName']
admission_notes_df = admission_notes_df[keepColumns]
admission_notes_df.head()

# rename columns
rename_dict = { 
    'CreateDate': 'NoteDate', 
    'DeidentifiedName': 'NoteTitle', 
}

admission_notes_df = admission_notes_df.rename(columns=rename_dict)

# change date column from str to datetime
admission_notes_df.loc[:, 'NoteDate'] = pd.to_datetime(admission_notes_df['NoteDate'], format='%Y%m%d')
admission_notes_df.head()

# read in icd filtered df
icd_filter_df = pd.read_csv('1_icd_neg_df.csv')
# change column from str to datetime
icd_filter_df['AdmissionDate'] = pd.to_datetime(icd_filter_df['AdmissionDate'])

# merge based on patient id
ICD_plus_AN_df = pd.merge(icd_filter_df, admission_notes_df, on='BDSPPatientID', how='inner')
ICD_plus_AN_df.head()


# filter for DateICD +/- 1 month of DateNote

from datetime import timedelta


time_window = timedelta(days=30)

# Filter rows where DateICD is within 1 month before or after DateNote
filtered_df = ICD_plus_AN_df[abs(ICD_plus_AN_df['AdmissionDate'] - ICD_plus_AN_df['NoteDate']) <= time_window]

# # Filter rows where DateICD is within 1 month after DateNote
# filtered_df = ICD_plus_DS_df[(ICD_plus_DS_df['DateICD'] >= ICD_plus_DS_df['DateNote']) & (ICD_plus_DS_df['DateICD'] <= ICD_plus_DS_df['DateNote'] + time_window)]

print(len(filtered_df))
filtered_df.head()


# RANDOMIZE order of df, 25 seed for reproducability
filtered_df = filtered_df.sample(frac=1, random_state=25).reset_index(drop=True)
filtered_df.head()

#filter to only include unique patients
filtered_df = filtered_df.drop_duplicates(subset='BDSPPatientID', keep='first')
print(len(filtered_df))

# save as csv
filtered_df.to_csv('2_admission_notes_df.csv', index=False)
