==> newFlowCohortExtractionPipeline/BIDMC/icd_neg/4_unique_patients.ipynb <==
import pandas as pd
from tqdm import tqdm

cpt_filtered_chunk = pd.read_csv('3_cpt_filtered_chunk2.csv')
cpt_filtered_chunk.head()


unique_patients = cpt_filtered_chunk.drop_duplicates(subset='BDSPPatientID', keep='first')
print(len(unique_patients))

def check_word_count(text, threshold):
    word_count = len(text.split())
    return word_count > threshold

threshold = 500
unique_patients['valid'] = unique_patients['text'].apply(lambda x: check_word_count(x, threshold))
unique_patients.head()

numValid = sum(unique_patients['valid'])
print(numValid)

false_indeces = unique_patients.index[~unique_patients['valid']].tolist()
print(false_indeces)

valid_notes = unique_patients.drop(index=false_indeces)
print(len(valid_notes))
valid_notes.head()

BIDMC_final_icd_neg = valid_notes
print(len(BIDMC_final_icd_neg))
BIDMC_final_icd_neg.head()

BIDMC_final_icd_neg = BIDMC_final_icd_neg.drop(columns=['valid'])
BIDMC_final_icd_neg.head()

BIDMC_final_icd_neg.to_csv('BIDMC_final_icd_neg.csv', index=False, header=True, sep=',', na_rep='NA')
